chore: we want to change the behavior of the APIClient, so we will split the packages
