+++ httpclient/httpclient_tools_test.go
@@ -1,0 +1,17 @@
+package httpclient
+
+import (
+	"testing"
+)
+
+func TestSwitchDebugLevel(t *testing.T) {
+
+}
+
+func TestConfigLog(t *testing.T) {
+
+}
+
+func TestRequestPath(t *testing.T) {
+
+}
